+++ nengo_bones/config.py
@@ -1,9 +1,46 @@
 """Handles the processing of nengo-bones configuration settings."""
 
 from collections import OrderedDict
+import datetime
 import os
 
 import yaml
+
+
+def check_list(cfg, key):
+    """
+    Verify that config value is a list.
+
+    This is designed to catch the common error of specifying
+
+    .. code-block:: yaml
+
+        option:
+            value
+
+    instead of
+
+    .. code-block:: yaml
+
+        option:
+            - value
+
+    Parameters
+    ----------
+    cfg : dict
+        Configuration options being checked.
+    key: str
+        Name of configuration value to be checked.
+
+    Raises
+    ------
+    TypeError
+        If ``key`` is in ``cfg`` and its value is not a list.
+    """
+    if key in cfg and not isinstance(cfg[key], list):
+        raise TypeError(
+            "%s should be a list, found '%s'; did you forget "
+            "to add '-' before each entry?" % (key, cfg[key]))
 
 
 def find_config():
@@ -30,21 +67,55 @@ def fill_defaults(config):
     config : dict
         Dictionary containing configuration values.
     """
-    if "travis_yml" in config:
-        config["travis_yml"].setdefault("python", "3.6")
-        config["travis_yml"].setdefault("global_vars", {})
-        config["travis_yml"].setdefault("pypi_user", None)
-        config["travis_yml"].setdefault("deploy_dists", ["sdist"])
-        config["travis_yml"].setdefault("bones_install", "nengo-bones")
 
-        for job in config["travis_yml"]["jobs"]:
+    config.setdefault("author", "Applied Brain Research")
+    config.setdefault("author_email", "info@appliedbrainresearch.com")
+    config.setdefault("copyright_start", datetime.datetime.now().year)
+    config.setdefault("copyright_end", datetime.datetime.now().year)
+
+    if "travis_yml" in config:
+        cfg = config["travis_yml"]
+        cfg.setdefault("python", "3.6")
+        cfg.setdefault("global_vars", OrderedDict())
+        cfg.setdefault("pypi_user", None)
+        cfg.setdefault("deploy_dists", ["sdist"])
+        cfg.setdefault("bones_install", "nengo-bones")
+
+        for job in cfg["jobs"]:
             if job.get("script", "").startswith("docs"):
                 job.setdefault("apt_install", ["pandoc"])
 
     if "codecov_yml" in config:
-        config["codecov_yml"].setdefault("skip_appveyor", True)
-        config["codecov_yml"].setdefault("abs_target", "auto")
-        config["codecov_yml"].setdefault("diff_target", "100%")
+        cfg = config["codecov_yml"]
+        cfg.setdefault("skip_appveyor", True)
+        cfg.setdefault("abs_target", "auto")
+        cfg.setdefault("diff_target", "100%")
+
+    if "license_rst" in config:
+        cfg = config["license_rst"]
+        cfg.setdefault("type", "nengo")
+
+    if "setup_py" in config:
+        cfg = config["setup_py"]
+        cfg.setdefault("license", "Free for non-commercial use")
+        cfg.setdefault("python_requires", ">=3.5")
+        cfg.setdefault("include_package_data", False)
+        cfg.setdefault(
+            "url",
+            "https://www.nengo.ai/%s" % config["pkg_name"].replace("_", "-"))
+
+    if "setup_cfg" in config:
+        cfg = config["setup_cfg"]
+        cfg.setdefault("pytest", OrderedDict())
+        cfg.setdefault("pylint", OrderedDict())
+        cfg.setdefault("flake8", OrderedDict())
+        cfg.setdefault("coverage", OrderedDict())
+        cfg["pytest"].setdefault("xfail_strict", False)
+        cfg["pytest"].setdefault("addopts", ["-p nengo.tests.options"])
+
+    if "docs_conf_py" in config:
+        cfg = config["docs_conf_py"]
+        cfg.setdefault("nengo_logo", "general-full-light.svg")
 
 
 def validate_config(config):
@@ -56,17 +127,19 @@ def validate_config(config):
     config : dict
         Dictionary containing configuration values.
     """
-    mandatory = ["pkg_name", "repo_name"]
-    if "travis_yml" in config:
-        mandatory.append("travis_yml.jobs")
+    mandatory = ["project_name", "pkg_name", "repo_name", "travis_yml.jobs"]
 
     for entry in mandatory:
         tmp = config
-        for key in entry.split("."):
+        for i, key in enumerate(entry.split(".")):
             try:
                 tmp = tmp[key]
             except KeyError:
-                raise KeyError("Config file must define %s" % entry)
+                if "." in entry and i == 0:
+                    # if the toplevel isn't defined, ignore this
+                    break
+                else:
+                    raise KeyError("Config file must define %s" % entry)
 
     if "ci_scripts" in config:
         for ci_config in config["ci_scripts"]:
@@ -94,11 +167,7 @@ def validate_ci_config(ci_config):
         # pip_install:
         #   - dependency
         for key in ("pip_install", "pre_commands", "post_commands"):
-            if not isinstance(ci_config[key], list):
-                raise TypeError(
-                    "%s should be a list, found '%s'; did you forget "
-                    "to add '-' before each entry?" % (
-                        key, ci_config[key]))
+            check_list(ci_config, key)
     except KeyError:
         pass