+++ nengo_bones/scripts/generate_bones.py
@@ -10,6 +10,41 @@ import jinja2
 import nengo_bones
 
 
+def render_template(ctx, cfg_key, output_file, template_file=None, **kwargs):
+    """
+    Render a template to file.
+
+    Parameters
+    ----------
+    ctx : `click.Context`
+        CLI context, containing information specified upstream.
+    cfg_key : str
+        The heading for the section in the config file containing config
+        options specific to the template being rendered.
+    output_file : str
+        Filename for the rendered output file.
+    template_file
+        Filename for the input template file (default is
+        ``output_file + ".template"``)
+    kwargs : dict
+        Will be passed on to the ``render`` function.
+    """
+    if template_file is None:
+        template_file = output_file + ".template"
+
+    template = ctx.obj["env"].get_template(template_file)
+    output_file = os.path.join(ctx.obj["output_dir"], output_file)
+    with open(output_file, "w") as f:
+        f.write(template.render(
+            version=nengo_bones.__version__,
+            **kwargs,
+            # pass in the top-level config options as well
+            # TODO: separate "top-level" config into its own section?
+            **ctx.obj["config"],
+            **ctx.obj["config"][cfg_key],
+        ))
+
+
 @click.group(invoke_without_command=True)
 @click.option("--conf-file", default=None, help="Filepath for config file")
 @click.option("--output-dir", default=".", help="Output directory for scripts")
@@ -33,6 +68,7 @@ def main(ctx, conf_file, output_dir, template_dir):
     env = jinja2.Environment(
         loader=jinja2.FileSystemLoader(template_dirs),
         trim_blocks=True, lstrip_blocks=True, keep_trailing_newline=True)
+    env.filters["rstrip"] = lambda s, chars: s.rstrip(chars)
 
     if not os.path.exists(output_dir):
         os.makedirs(os.path.normpath(output_dir))
@@ -41,9 +77,19 @@ def main(ctx, conf_file, output_dir, template_dir):
     ctx.obj["config"] = config
     ctx.obj["output_dir"] = output_dir
 
+    def check_cfg(name):
+        name = name.replace("-", "_")
+        if name not in config:
+            print("No config entry detected for %s, skipping" % name)
+            return False
+        return True
+
     if ctx.invoked_subcommand is None:
-        for script in (travis_yml, codecov_yml):
-            ctx.invoke(script)
+        for cfg_name in nengo_bones.all_templated_files.values():
+            if check_cfg(cfg_name):
+                ctx.invoke(globals()[cfg_name])
+    elif not check_cfg(ctx.invoked_subcommand):
+        exit()
 
 
 @main.command()
@@ -95,13 +141,87 @@ def travis_yml(ctx):
 def codecov_yml(ctx):
     """Generate codecov config file."""
 
+    render_template(ctx, "codecov_yml", ".codecov.yml")
+
+
+@main.command()
+@click.pass_context
+def license_rst(ctx):
+    """Generate LICENSE.rst file."""
+
+    render_template(ctx, "license_rst", "LICENSE.rst")
+
+
+@main.command()
+@click.pass_context
+def contributing_rst(ctx):
+    """Generate CONTRIBUTING.rst file."""
+
+    render_template(ctx, "contributing_rst", "CONTRIBUTING.rst")
+
+
+@main.command()
+@click.pass_context
+def contributors_rst(ctx):
+    """Generate CONTRIBUTORS.rst file."""
+
+    render_template(ctx, "contributors_rst", "CONTRIBUTORS.rst")
+
+
+@main.command()
+@click.pass_context
+def manifest_in(ctx):
+    """Generate MANIFEST.in file."""
+
+    render_template(ctx, "manifest_in", "MANIFEST.in",
+                    custom=ctx.obj["config"]["manifest_in"])
+
+
+@main.command()
+@click.pass_context
+def setup_py(ctx):
+    """Generate setup.py file."""
+
     config = ctx.obj["config"]
 
-    template = ctx.obj["env"].get_template(".codecov.yml.template")
-    with open(os.path.join(ctx.obj["output_dir"], ".codecov.yml"), "w") as f:
-        f.write(template.render(
-            version=nengo_bones.__version__,
-            **config["codecov_yml"]))
+    setup_config = config["setup_py"]
+
+    extras = OrderedDict()
+    extra_types = {
+        "classifiers": "list",
+        "entry_points": "dict",
+        "package_data": "dict",
+    }
+    # we iterate over setup_config (rather than extra_types) so that the
+    # order in setup_config is preserved
+    for key in list(setup_config.keys()):
+        if key in extra_types:
+            extras[key] = (extra_types[key], setup_config.pop(key))
+
+    render_template(ctx, "setup_py", "setup.py",
+                    extras=extras)
+
+
+@main.command()
+@click.pass_context
+def setup_cfg(ctx):
+    """Generate setup.cfg file."""
+
+    render_template(ctx, "setup_cfg", "setup.cfg")
+
+
+@main.command()
+@click.pass_context
+def docs_conf_py(ctx):
+    """Generate docs/conf.py file."""
+
+    output_dir = os.path.join(ctx.obj["output_dir"], "docs")
+
+    if not os.path.exists(output_dir):
+        os.makedirs(output_dir)
+
+    render_template(ctx, "docs_conf_py", "docs/conf.py",
+                    template_file="docs_conf.py.template")
 
 
 if __name__ == "__main__":